+++ tests/config.rs
@@ -1,3 +1,4 @@
+use dook::downloads_policy::get_downloads_policy;
 use dook::loader;
 use dook::{Config, QueryCompiler};
 
@@ -5,10 +6,13 @@ use dook::{Config, QueryCompiler};
 fn default_patterns_are_loadable() {
     let target_dir = std::path::PathBuf::from(env!("CARGO_TARGET_TMPDIR"));
     eprintln!("CARGO_TARGET_TMPDIR is {:?}", target_dir);
-    let mut language_loader =
-        loader::Loader::new(target_dir.clone(), Some(target_dir.clone()), false).expect(
-            "should have called tree_sitter_loader::Loader::with_parser_lib_path(), not new()",
-        );
+    let downloads_policy = get_downloads_policy();
+    let mut language_loader = loader::Loader::new(
+        target_dir.clone(),
+        Some(target_dir.clone()),
+        downloads_policy,
+    )
+    .expect("should have called tree_sitter_loader::Loader::with_parser_lib_path(), not new()");
     let default_config = Config::load_default();
     let mut query_compiler = QueryCompiler::new(&default_config);
     for language_name in default_config.configured_languages() {